requirements.txt
Flask
pymongo
dnspython
gunicorn
flask-cors
python-dateutil
python-dotenv